+++ algorithm/transposition_table.c
@@ -27,7 +27,7 @@ static bool TT_lookup(uint64_t hash, TT_Entry *entry)
     return true;
 }
 
-void TT_store(uint64_t hash, uint8_t depth, int32_t score,
+void TT_store(uint64_t hash, uint8_t depth, double score,
               Result result, TT_Entry_Type type, uint16_t move)
 {
     size_t index = hash & TT_MASK;
